test(ios): move #4394 caret focus regression into caret.ts
Fold the caretFocus.ts regression spec back into the Caret describe block
in caret.ts and delete the standalone file. No behavior change; the spec
still asserts the keyboard stays down on this fixed branch.
